+++ apis/clusterresources/v1beta1/azurevnetpeering_webhook.go
@@ -18,6 +18,7 @@ package v1beta1
 
 import (
 	"fmt"
+	"regexp"
 
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -103,5 +104,27 @@ func (r *AzureVNetPeering) ValidateDelete() error {
 	azurevnetpeeringlog.Info("validate delete", "name", r.Name)
 
 	// TODO(user): fill in your validation logic upon object deletion.
+	return nil
+}
+
+func (azure *AzureVNetPeeringSpec) Validate() error {
+	dataCentreIDMatched, err := regexp.Match(models.UUIDStringRegExp, []byte(azure.DataCentreID))
+	if err != nil {
+		return err
+	}
+	if !dataCentreIDMatched {
+		return fmt.Errorf("data centre ID is a UUID formated string. It must fit the pattern: %s", models.UUIDStringRegExp)
+	}
+
+	for _, subnet := range azure.PeerSubnets {
+		peerSubnetMatched, err := regexp.Match(models.PeerSubnetsRegExp, []byte(subnet))
+		if err != nil {
+			return err
+		}
+		if !peerSubnetMatched {
+			return fmt.Errorf("the provided CIDR: %s must contain four dot separated parts and form the Private IP address. All bits in the host part of the CIDR must be 0. Suffix must be between 16-28", subnet)
+		}
+	}
+
 	return nil
 }
