+++ apis/clusterresources/v1beta1/exclusionwindow_types.go
@@ -72,17 +72,3 @@ func (r *ExclusionWindow) NewPatch() client.Patch {
 	old := r.DeepCopy()
 	return client.MergeFrom(old)
 }
-
-func (e *ExclusionWindowSpec) validateUpdate(old ExclusionWindowSpec) bool {
-	if e.DayOfWeek != old.DayOfWeek ||
-		e.ClusterID != old.ClusterID ||
-		e.DurationInHours != old.DurationInHours ||
-		e.StartHour != old.StartHour ||
-		(e.ClusterRef != nil && old.ClusterRef == nil) ||
-		(e.ClusterRef == nil && old.ClusterRef != nil) ||
-		(e.ClusterRef != nil && *e.ClusterRef != *old.ClusterRef) {
-		return false
-	}
-
-	return true
-}
