+++ apis/clusterresources/v1beta1/gcpvpcpeering_webhook.go
@@ -18,6 +18,7 @@ package v1beta1
 
 import (
 	"fmt"
+	"regexp"
 
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -99,5 +100,27 @@ func (r *GCPVPCPeering) ValidateDelete() error {
 	gcpvpcpeeringlog.Info("validate delete", "name", r.Name)
 
 	// TODO(user): fill in your validation logic upon object deletion.
+	return nil
+}
+
+func (gcp *GCPVPCPeeringSpec) Validate() error {
+	dataCentreIDMatched, err := regexp.Match(models.UUIDStringRegExp, []byte(gcp.DataCentreID))
+	if err != nil {
+		return err
+	}
+	if !dataCentreIDMatched {
+		return fmt.Errorf("data centre ID is a UUID formated string. It must fit the pattern: %s", models.UUIDStringRegExp)
+	}
+
+	for _, subnet := range gcp.PeerSubnets {
+		peerSubnetMatched, err := regexp.Match(models.PeerSubnetsRegExp, []byte(subnet))
+		if err != nil {
+			return err
+		}
+		if !peerSubnetMatched {
+			return fmt.Errorf("the provided CIDR: %s must contain four dot separated parts and form the Private IP address. All bits in the host part of the CIDR must be 0. Suffix must be between 16-28", subnet)
+		}
+	}
+
 	return nil
 }
