+++ apis/clusters/v1beta1/kafkaconnect_webhook.go
@@ -224,3 +224,144 @@ func (kcv *kafkaConnectValidator) ValidateDelete(ctx context.Context, obj runtim
 	// TODO(user): fill in your validation logic upon object deletion.
 	return nil
 }
+
+type immutableKafkaConnectFields struct {
+	immutableCluster
+}
+
+type immutableKafkaConnectDCFields struct {
+	immutableDC       immutableDC
+	ReplicationFactor int
+}
+
+func (kc *KafkaConnectSpec) newImmutableFields() *immutableKafkaConnectFields {
+	return &immutableKafkaConnectFields{
+		immutableCluster: kc.Cluster.newImmutableFields(),
+	}
+}
+
+func (kc *KafkaConnectSpec) validateUpdate(oldSpec KafkaConnectSpec) error {
+	newImmutableFields := kc.newImmutableFields()
+	oldImmutableFields := oldSpec.newImmutableFields()
+
+	if *newImmutableFields != *oldImmutableFields {
+		return fmt.Errorf("cannot update immutable fields: old spec: %+v: new spec: %+v", oldSpec, kc)
+	}
+
+	err := kc.validateImmutableDataCentresFieldsUpdate(oldSpec)
+	if err != nil {
+		return err
+	}
+
+	err = kc.validateImmutableTargetClusterFieldsUpdate(kc.TargetCluster, oldSpec.TargetCluster)
+	if err != nil {
+		return err
+	}
+
+	err = validateTwoFactorDelete(kc.TwoFactorDelete, oldSpec.TwoFactorDelete)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (kc *KafkaConnectSpec) validateImmutableDataCentresFieldsUpdate(oldSpec KafkaConnectSpec) error {
+	if len(kc.DataCentres) != len(oldSpec.DataCentres) {
+		return models.ErrImmutableDataCentresNumber
+	}
+
+	for i, newDC := range kc.DataCentres {
+		oldDC := oldSpec.DataCentres[i]
+		newDCImmutableFields := newDC.newImmutableFields()
+		oldDCImmutableFields := oldDC.newImmutableFields()
+
+		if *newDCImmutableFields != *oldDCImmutableFields {
+			return fmt.Errorf("cannot update immutable data centre fields: new spec: %v: old spec: %v", newDCImmutableFields, oldDCImmutableFields)
+		}
+
+		err := newDC.validateImmutableCloudProviderSettingsUpdate(oldDC.CloudProviderSettings)
+		if err != nil {
+			return err
+		}
+
+		err = validateTagsUpdate(newDC.Tags, oldDC.Tags)
+		if err != nil {
+			return err
+		}
+
+		if newDC.NodesNumber < oldDC.NodesNumber {
+			return fmt.Errorf("deleting nodes is not supported. Number of nodes must be greater than: %v", oldDC.NodesNumber)
+		}
+
+		if ((newDC.NodesNumber*newDC.ReplicationFactor)/newDC.ReplicationFactor)%newDC.ReplicationFactor != 0 {
+			return fmt.Errorf("number of nodes must be a multiple of replication factor: %v", newDC.ReplicationFactor)
+		}
+	}
+
+	return nil
+}
+
+func (kdc *KafkaConnectDataCentre) newImmutableFields() *immutableKafkaConnectDCFields {
+	return &immutableKafkaConnectDCFields{
+		immutableDC: immutableDC{
+			Name:                kdc.Name,
+			Region:              kdc.Region,
+			CloudProvider:       kdc.CloudProvider,
+			ProviderAccountName: kdc.ProviderAccountName,
+			Network:             kdc.Network,
+		},
+		ReplicationFactor: kdc.ReplicationFactor,
+	}
+}
+
+func (kc *KafkaConnectSpec) validateImmutableTargetClusterFieldsUpdate(new, old []*TargetCluster) error {
+	if len(new) == 0 && len(old) == 0 {
+		return models.ErrImmutableTargetCluster
+	}
+
+	if len(old) != len(new) {
+		return models.ErrImmutableTargetCluster
+	}
+
+	for _, index := range new {
+		for _, elem := range old {
+			err := validateImmutableExternalClusterFields(index, elem)
+			if err != nil {
+				return err
+			}
+
+			err = validateImmutableManagedClusterFields(index, elem)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func validateImmutableExternalClusterFields(new, old *TargetCluster) error {
+	for _, index := range new.ExternalCluster {
+		for _, elem := range old.ExternalCluster {
+			if *index != *elem {
+				return models.ErrImmutableExternalCluster
+			}
+		}
+	}
+	return nil
+}
+
+func validateImmutableManagedClusterFields(new, old *TargetCluster) error {
+	for _, nm := range new.ManagedCluster {
+		for _, om := range old.ManagedCluster {
+			if nm.TargetKafkaClusterID != om.TargetKafkaClusterID ||
+				nm.KafkaConnectVPCType != om.KafkaConnectVPCType ||
+				(nm.ClusterRef != nil && om.ClusterRef == nil) ||
+				(nm.ClusterRef == nil && om.ClusterRef != nil) ||
+				(nm.ClusterRef != nil && *nm.ClusterRef != *om.ClusterRef) {
+				return models.ErrImmutableManagedCluster
+			}
+		}
+	}
+	return nil
+}
